Standard Eclipse/STS Java Clean-up Settings, sorted
These are the same settings as before, just in a different order.

Whenever you touch any of the settings, even if you then undo what
you've done, the settings seem to get exported in a very different
order.
This makes it hard to see subsequent individual changes. So we save
this sorted order first, before making any changes to the settings.