feat: add corresponding Eclipse default files
These are the default settings from the "Eclipse [built in]" version,
but sorted.
Our modified settings files started out the same as these, so the
changes can easily be seen in the version control history, but it's
sometimes useful to have the original settings available in a separate
file. So they can be compated easily.