style(std): consistent use of lists  for sections
To be consistent with other parts of the document, don't use Markdown
headings to define/show different sections of Eclipse settings, use
list items, and indent the individual settings an extra level within.
We were ending up with heading level 5, which is too deep, and too
dense, and is not really related to our document's structure but to the
structure of the dialogs in Eclipse, so was overkill.