style: legacy Javadoc @param indent settings
Update the style guide to control how @param and @throws Javadoc
comments are indented: new line, and align with tag. These settings
produce formatting that matches the existing code base, which was
created with default Eclipse settings. So it appears that Eclipse have
changed what their defaults are, to default to no new lines.

We want to match our existing code. That style also more closely
matches the default way the Javadocs are rendered, for example in the
Eclipse Javadoc "hover" display.